+++ src-d/rules_python/python/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2019 The Bazel Authors. All rights reserved.
+# Copyright 2024 The Bazel Authors. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,20 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-load("//python:py_library.bzl", "py_library")
-
-filegroup(
-    name = "distribution",
-    srcs = glob(["**"]),
-    visibility = ["//python:__pkg__"],
-)
-
-py_library(
-    name = "runfiles",
-    srcs = [
-        "__init__.py",
-        "runfiles.py",
-    ],
-    visibility = ["//visibility:public"],
-    deps = ["//src-d/rules_python/python/runfiles"],
-)
+# This happens when:
+# * bzlmod is enabled
+# * `import python.runfiles` is run
+if __name__ == "python":
+    import rules_python.python
+    import sys
+    sys.modules[__name__] = rules_python.python