+++ CHANGELOG.md
@@ -37,6 +37,9 @@ A brief description of the categories of changes:
   platform-specific content in `MODULE.bazel.lock` files; Follow
   [#1643](https://github.com/bazelbuild/rules_python/issues/1643) for removing
   platform-specific content in `MODULE.bazel.lock` files.
+* (bzlmod runfiles) The runfiles library can now be imported using
+  `import rules_python.python.runfiles`, as is possible with wORKSPACE builds.
+  ([#1679](https://github.com/bazelbuild/rules_python/issues/1679))
 * (wheel) The stamp variables inside the distribution name are no longer
   lower-cased when normalizing under PEP440 conventions.